+++ CHANGELOG.md
@@ -2,5 +2,6 @@
 
 ## v1.2.0
 - added sound effects and a more advanced config system
+- added selecting to cards + random button
 
 wearsnomask guess who [V1.1.0 and under](https://github.com/wearsnomask/utdr-guess-who/blob/main/CHANGELOG.md)